Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
joblib==1.3.2 
pydantic-settings==2.0.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
